+++ src/embeds/pokemonEmbeds.js
@@ -369,13 +369,13 @@ const buildPokemonEmbed = (
   }
 
   let sixthField = {
-    name: "Shiny",
+    name: "✨ Shiny",
     value: pokemon.shiny ? "True" : "False",
     inline: true,
   };
   if (originalOwnerId) {
     sixthField = {
-      name: "Original Owner",
+      name: "👑 Original Owner",
       value: `<@${originalOwnerId}>`,
       inline: true,
     };
@@ -408,7 +408,9 @@ const buildPokemonEmbed = (
   embed.setDescription(
     `${pokemon.shiny ? "✨" : ""}**[Lv. ${pokemon.level}]** ${
       speciesData.name
-    } (#${pokemon.speciesId})\n${linebreakString(speciesData.description, 50)}`
+    } (#${pokemon.speciesId})\n${buildBlockQuoteString(
+      speciesData.description
+    )}`
   );
   embed.setColor(rarityConfig[speciesData.rarity].color);
 
@@ -420,32 +422,32 @@ const buildPokemonEmbed = (
         : "";
 
     embed.addFields(
-      { name: "Type", value: typeString, inline: true },
+      { name: "🏷️ Type", value: typeString, inline: true },
       {
-        name: "Ability",
+        name: "💫 Ability",
         value: getAbilityName(pokemon.abilityId),
         inline: true,
       },
       {
-        name: "Nature",
+        name: "🧬 Nature",
         value: `${natureConfig[pokemon.natureId].name} (${
           natureConfig[pokemon.natureId].description
         })`,
         inline: true,
       },
-      { name: "Rarity", value: speciesData.rarity, inline: true },
+      { name: "💎 Rarity", value: speciesData.rarity, inline: true },
       {
-        name: "Date Caught",
+        name: "🗓️ Date Caught",
         value: new Date(pokemon.dateAcquired).toLocaleDateString(),
         inline: true,
       },
       sixthField,
       {
-        name: "Stats (Stat|IVs|EVs)",
+        name: "💪 Stats (Stat|IVs|EVs)",
         value: `${statString}${heldItemString}`, // TODO: can't find a better place to put this lol
         inline: false,
       },
-      { name: "Level Progress", value: progressBar, inline: false }
+      { name: "⏳ Level Progress", value: progressBar, inline: false }
     );
 
     footerHelp.push("/train <id> to train this Pokemon");
@@ -474,7 +476,9 @@ const buildPokemonEmbed = (
     const abilityData = getAbility(pokemon.abilityId);
     embed.addFields({
       name: `Ability: ${getAbilityName(pokemon.abilityId)}`,
-      value: abilityData ? abilityData.description : "Not yet implemented!",
+      value: abilityData
+        ? buildBlockQuoteString(abilityData.description)
+        : "Not yet implemented!",
       inline: false,
     });
 
@@ -482,7 +486,9 @@ const buildPokemonEmbed = (
     if (pokemon.heldItemId) {
       embed.addFields({
         name: `Held Item: ${getItemDisplay(pokemon.heldItemId)}`,
-        value: backpackItemConfig[pokemon.heldItemId].description,
+        value: buildBlockQuoteString(
+          backpackItemConfig[pokemon.heldItemId].description
+        ),
         inline: false,
       });
     }
@@ -514,7 +520,7 @@ const buildPokemonEmbed = (
     // add stat boost field
     if (oldPokemon) {
       embed.addFields({
-        name: "Stat Boost",
+        name: "⬆️ Stat Boost",
         value: buildBoostString(oldPokemon, pokemon),
         inline: false,
       });