+++ changelog.md
@@ -5,6 +5,7 @@
 - Improve UI
   - Improve look of fields, use inline code when appropriate, use quotes to indent, use emojis!
   - Also put more emojis in buttons
+  - Move, ability, and held item description formatting?
 - Put tutorial button in upsells
   - Make the tutorial commands big asf
   - Can add tips as footers